feat(openfeedback-m3): support ios target.
