+++ openfeedback-m3/build.gradle.kts
@@ -1,7 +1,7 @@
-
 plugins {
     id("io.openfeedback.plugins.lib.multiplatform")
     id("io.openfeedback.plugins.publishing")
+    kotlin("native.cocoapods")
     alias(libs.plugins.jetbrains.compose)
 }
 
@@ -14,7 +14,31 @@ openfeedback {
 }
 
 kotlin {
+    applyDefaultHierarchyTemplate()
+
     androidTarget()
+
+    listOf(
+        iosX64(),
+        iosArm64(),
+        iosSimulatorArm64()
+    ).forEach { iosTarget ->
+        iosTarget.binaries.framework {
+            baseName = "OpenFeedbackM3Kit"
+            isStatic = true
+        }
+    }
+
+    cocoapods {
+        version = "1.0"
+        ios.deploymentTarget = "14.1"
+        pod("FirebaseAuth") {
+            linkOnly = true
+        }
+        pod("FirebaseFirestore") {
+            linkOnly = true
+        }
+    }
 
     sourceSets {
         val commonMain by getting {
@@ -27,9 +51,12 @@ kotlin {
 
                 implementation(compose.material3)
                 implementation(compose.ui)
+            }
+        }
+        val androidMain by getting {
+            dependencies {
                 implementation(compose.uiTooling)
             }
         }
-        val androidMain by getting
     }
 }
