+++ openfeedback-m3/src/androidMain/kotlin/io/openfeedback/m3/LoadingPreview.kt
@@ -1,0 +1,10 @@
+package io.openfeedback.m3
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.tooling.preview.Preview
+
+@Preview
+@Composable
+internal fun LoadingPreview() {
+    Loading()
+}
