+++ openfeedback-m3/src/commonMain/kotlin/io/openfeedback/m3/CommentInput.kt
@@ -6,13 +6,11 @@ import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.outlined.Send
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextField
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.input.ImeAction
-import androidx.compose.ui.tooling.preview.Preview
 import dev.icerock.moko.resources.compose.stringResource
 import io.openfeedback.MR
 
@@ -42,16 +40,4 @@ fun CommentInput(
         keyboardOptions = KeyboardOptions(imeAction = ImeAction.Done),
         maxLines = 5
     )
-}
-
-@Preview
-@Composable
-private fun CommentInputPreview() {
-    MaterialTheme {
-        CommentInput(
-            value = "My comment",
-            onValueChange = {},
-            onSubmit = {}
-        )
-    }
 }
